Map all C++ integer types (int, long, short) to Rust i64
Avoids type mismatch errors when mixing int and long long in operations.
C++ allows implicit widening between integer types, Rust requires explicit
casts. Mapping everything to i64 produces compilable Rust code for the
MVP (LeetCode-style algorithm) use case.

+++ src/codegen/rust.rs
@@ -1204,9 +1204,9 @@ fn map_cpp_type_to_rust(name: &str) -> String {
         "ptrdiff_t" => "isize".to_string(),
         "float" => "f32".to_string(),
         "double" => "f64".to_string(),
-        "int" => "i32".to_string(),
+        "int" => "i64".to_string(),
         "long" => "i64".to_string(),
-        "short" => "i16".to_string(),
+        "short" => "i64".to_string(),
         other => other.to_string(), // fallback
     }
 }
